requirements.txt
discord.py==2.7.0
python-dotenv==1.0.0
